+++ index.ts
@@ -3,33 +3,8 @@ const Typespess = require("./code/game/server.js");
 const read_config = require("./code/config.js");
 const World = require("./code/game/world.js");
 const Database = require("./code/database.js");
-const {app, BrowserWindow} = require('electron');
 
 const {URLSearchParams} = require("url");
-
-function createWindow () {
-	const win = new BrowserWindow({
-		width: 800,
-		height: 600,
-		webPreferences: {nodeIntegration: true}
-	});
-
-	win.loadFile('./resources/index.html');
-}
-
-app.whenReady().then(createWindow);
-
-app.on('window-all-closed', () => {
-	if (process.platform !== 'darwin') {
-		app.quit();
-	}
-});
-
-app.on('activate', () => {
-	if (BrowserWindow.getAllWindows().length === 0) {
-		createWindow();
-	}
-});
 
 console.info("SERVER: Loading game...");
